refactor(server): split up index.ts
closes #98
