+++ package.json
@@ -132,6 +132,7 @@
     "@storybook/vue": "3.3.15",
     "@types/animejs": "2.0.0",
     "@types/axios-mock-adapter": "1.10.0",
+    "@types/compression": "0.0.36",
     "@types/cookie-parser": "1.4.1",
     "@types/deepmerge": "1.3.3",
     "@types/express": "4.11.1",
